Adjusted tooltip to stay on screen when
